+++ migrations/token/metadata/TokenMetadataResolverRandomizedReveal@3.0_deploy.js
@@ -6,8 +6,8 @@ module.exports = Contract_deploy('TokenMetadataResolverRandomizedReveal@3.0', {
   contract: 'TokenMetadataResolverRandomizedReveal',
   importPath: 'node_modules/@animoca/ethereum-contracts-3.0/artifacts',
   args: [
-    {name: 'Chainlink_LinkToken', value: getContractAddressOrFallback('Chainlink_LinkToken', ethers.constants.AddressZero)},
-    {name: 'Chainlink_VRFV2Wrapper', value: getContractAddressOrFallback('Chainlink_VRFV2Wrapper', ethers.constants.AddressZero)},
+    {name: 'Chainlink_LinkToken', value: getContractAddressOrFallback('Chainlink_LinkToken', ethers.ZeroAddress)},
+    {name: 'Chainlink_VRFV2Wrapper', value: getContractAddressOrFallback('Chainlink_VRFV2Wrapper', ethers.ZeroAddress)},
   ],
   deterministicDeployment: true,
 });